+++ crates/lib/src/lib.rs
@@ -49,6 +49,9 @@ mod http;
 mod socket;
 mod utils;
 
+#[cfg(feature = "packet_capture")]
+pub mod capture;
+
 pub use errors::*;
 #[cfg(feature = "games")]
 pub use games::*;